app/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastai==2.7.15
fastapi==0.111.1
fastapi-cli==0.0.4
fastcore==1.5.54
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
pillow==10.4.0
preshed==3.0.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
sympy==1.13.1
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.7.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
